requirements.txt
Django==4.2
djangorestframework==3.14.0
Unidecode==1.3.6
psycopg2-binary==2.9.6
Pillow==9.5.0